pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fastfixupfinder"
version = "0.1.0"
description = "A tool to automatically identify fixup targets and create fixup commits"
authors = [{name = "Developer"}]
requires-python = ">=3.10"
dependencies = [
    "click>=8.0.0",
    "gitpython>=3.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=22.0.0",
    "flake8>=4.0.0",
    "mypy>=1.0.0",
]

[project.scripts]
fastfixupfinder = "fastfixupfinder.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["fastfixupfinder*"]

[tool.black]
line-length = 88
target-version = ['py310']

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true